+++ libraries/package-extractor/src/Utils.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { Text } from '@rushstack/node-core-library';
+import path from 'node:path';
+import { Path, Text } from '@rushstack/node-core-library';
 
 export function matchesWithStar(patternWithStar: string, input: string): boolean {
   // Map "@types/*" --> "^\@types\/.*$"
@@ -15,4 +16,41 @@ export function matchesWithStar(patternWithStar: string, input: string): boolean
   // eslint-disable-next-line @rushstack/security/no-unsafe-regexp
   const regExp: RegExp = new RegExp(pattern);
   return regExp.test(input);
+}
+
+export interface IRemapPathForTargetFolder {
+  sourcePath: string;
+  sourceRootFolder: string;
+  targetRootFolder: string;
+}
+
+/**
+ * Maps a file path under the provided {@link IRemapPathForTargetFolder.sourceRootFolder} to the provided
+ * {@link IExtractorOptions.targetRootFolder}.
+ *
+ * Example input: "C:\\MyRepo\\libraries\\my-lib"
+ * Example output: "C:\\MyRepo\\common\\deploy\\libraries\\my-lib"
+ */
+export function remapSourcePathForTargetFolder(options: IRemapPathForTargetFolder): string {
+  const { sourcePath, sourceRootFolder, targetRootFolder } = options;
+  const relativePath: string = path.relative(sourceRootFolder, sourcePath);
+  if (relativePath.startsWith('..')) {
+    throw new Error(`Source path "${sourcePath}" is not under "${sourceRootFolder}"`);
+  }
+  const absolutePathInTargetFolder: string = path.join(targetRootFolder, relativePath);
+  return absolutePathInTargetFolder;
+}
+
+/**
+ * Maps a file path under the provided folder path to the expected path format for the extractor metadata.
+ *
+ * Example input: "C:\\MyRepo\\libraries\\my-lib"
+ * Example output: "common/deploy/libraries/my-lib"
+ */
+export function remapPathForExtractorMetadata(folderPath: string, filePath: string): string {
+  const relativePath: string = path.relative(folderPath, filePath);
+  if (relativePath.startsWith('..')) {
+    throw new Error(`Path "${filePath}" is not under "${folderPath}"`);
+  }
+  return Path.convertToSlashes(relativePath);
 }