+++ libraries/debug-certificate-manager/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @rushstack/debug-certificate-manager
 
-This log was last generated on Tue, 10 Sep 2024 20:08:11 GMT and should not be manually modified.
+This log was last generated on Sat, 21 Sep 2024 00:10:27 GMT and should not be manually modified.
+
+## 1.4.0
+Sat, 21 Sep 2024 00:10:27 GMT
+
+### Minor changes
+
+- Add a `skipCertificateTrust` option to `CertificateManager.ensureCertificateAsync` that skips automatically trusting the generated certificate and untrusting an existing certificate with issues.
+
+## 1.3.66
+Fri, 13 Sep 2024 00:11:42 GMT
+
+_Version update only_
 
 ## 1.3.65
 Tue, 10 Sep 2024 20:08:11 GMT