+++ webpack/webpack4-localization-plugin/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/webpack4-localization-plugin
 
-This log was last generated on Tue, 10 Sep 2024 20:08:11 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Sep 2024 00:11:43 GMT and should not be manually modified.
+
+## 0.18.68
+Fri, 13 Sep 2024 00:11:43 GMT
+
+_Version update only_
 
 ## 0.18.67
 Tue, 10 Sep 2024 20:08:11 GMT